+++ repl_sessions/round_trip.clj
@@ -9,45 +9,62 @@
             [lambdaisland.regal.platform :as platform]))
 
 (gen/sample
-
  (s/gen :lambdaisland.regal/form))
 
-(defn join-strings [v]
-  (reduce (fn [v e]
-            (if (and (string? (last v)) (string? e))
-              (update v (dec (count v)) str e)
-              (conj v e)))
-          [] v))
-
-(defn normalize [form]
-  (cond
-    (vector? form)
-    (cond
-      (and (= 2 (count form)) (= :cat (first form)))
-      (recur (second form))
-
-      (keyword? (first form))
-      (let [joined (join-strings (mapv normalize form))]
-        (if (not= joined form)
-          (recur joined)
-          form))
-
-      :else
-      (mapv normalize form))
-
-    (char? form)
-    (str form)
-
-    :else
-    form))
+(defn round-trip? [form]
+  (= form (parse/parse (regal/regex form))))
 
 (check/quick-check
  100
- (prop/for-all [form (gen/fmap normalize (s/gen :lambdaisland.regal/form))]
-   (= form (parse/parse (regal/regex form))))
+ (prop/for-all*
+  [(gen/fmap regal/normalize (s/gen :lambdaisland.regal/form))]
+  round-trip?))
+
+(check/quick-check
+ 100
+ (prop/for-all [regal (s/gen ::regal/form)]
+   ;; test that generators generate valid regexes
+   (try
+     (regal/regex regal)
+     (catch Exception _
+       false))))
+
+(round-trip? [:cat "   " [:class "&& "]])
+
+
+(round-trip? [:class " " [" " "["]])
+
+(round-trip? [:ctrl "A"])
+
+(round-trip? [:class "   - "])
+
+(round-trip? [:alt "  " [:capture " " :escape]])
+
+(round-trip? :whitespace)
+
+(round-trip? [:? [:? "x"]])
+
+(round-trip? [:cat "  " [:class " " :non-whitespace]])
+
+(round-trip? [:cat "-" [:repeat [:repeat "x" 0] 0]])
+
+(dotimes [i 100]
+  (print i)
+  (time (doall (map round-trip? (gen/sample (s/gen ::regal/form) i)))))
+
+(re-find
+ #"-(?:x{2})?"
+ "-x"
+
  )
 
-(normalize [:cat "  "])
+(re-find #"ax??" "ax")
+
+(re-find #"[ ^]" "^")
+(re-find #"[$]" "$")
+(re-find #"[.]" "a")
+(re-find #"[a\-z]" "b")
+(re-find #"[a\-z]" "-")
 
 (regal/pattern [:not " " :whitespace])